Adjust to language changes immediately
Previously, if the language of the user interface was changed in the
settings, some texts immediately switched to the newly selected
language while others stayed in the old language.

Now the texts are no longer stored in the application, they are fetched
from the resources whenever they are needed. A side effect of this
change is that fewer data needs to be passed between the activity and
the fragment.