Always use left-to-right for user input
The input consists only of mathematical symbols, not of natural language
text.

Previously, using the app in a right-to-left language, such as Hebrew or
Arabic, started the input on the right-hand side of the input field, and
typing a '(' was shown as ')'.

Mentioned in https://github.com/atorch/probability_puzzles/issues/20.